Initial commit of server example
An example of what's required for sending GET requests on the LAN. Requests are sent to raspPi, a custom device handler, which can then execute any scripts Linux is capable of. Examples included in the file in myPCHandler will send magic packets over the LAN to wake up an ethernet-connected PC, or send a shutdown request to an EventGhost server running on a Windows PC. Integration with the Amazon Echo was achieved by tying virtual switches to the Alexa/SmartThings smartapp, then running specified routines based on switch state.